mcp-server/requirements.txt
mcp>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
python-dateutil>=2.8.0
